Moving types to a separate folder
