+++ components/atoms/button/button.tsx
@@ -1,15 +1,11 @@
 import { ReactNode, memo } from "react";
-import { ColorVariants } from "../../../constant/styles/colors";
-import { ButtonSizeVariants } from "../../../constant/styles/button";
 import {
-  PositionVariants,
-  SpaceVariants,
-} from "../../../constant/styles/common";
-
-type ColorKey = keyof typeof ColorVariants;
-type SizeKey = keyof typeof ButtonSizeVariants;
-type PositionKey = keyof typeof PositionVariants;
-type MarignSpaceKey = keyof typeof SpaceVariants;
+  ColorKey,
+  MarignSpaceKey,
+  PositionKey,
+  SizeKey,
+} from "@/types/variants";
+import { getButtonDynamicProps } from "../../../utils";
 
 export type ButtonProps = {
   children: ReactNode;
@@ -19,22 +15,6 @@ export type ButtonProps = {
   space?: MarignSpaceKey;
   onClick?: () => void;
   className?: string;
-};
-
-const getButtonDynamicProps = ({
-  bgColor,
-  size,
-  space = "default",
-  position = "center",
-}: Omit<ButtonProps, "children">) => {
-  return [
-    ColorVariants[bgColor].bg,
-    ButtonSizeVariants[size],
-    PositionVariants[position],
-    SpaceVariants[space],
-  ]
-    .filter((item) => !!item)
-    .join(" ");
 };
 
 export const Button = memo<ButtonProps>(
@@ -55,7 +35,7 @@ export const Button = memo<ButtonProps>(
           size,
           space,
           position,
-        })} ${className} flex justify-center items-center hover:scale-105 transition-all ease-in-out duration-300  hover:cursor-pointer text-md shadow-md  tracking-wider  rounded-md text-center`}
+        })} ${className} flex justify-center items-center hover:scale-105 transition-all ease-in-out duration-300  hover:cursor-pointer text-md text-white shadow-md  tracking-wider  rounded-md text-center`}
       >
         {children}
       </button>
